helper scala object created for google maps api related operations
