+++ app/controllers/GalleryController.scala
@@ -2,6 +2,7 @@ package controllers
 
 import javax.inject.Inject
 import controllers.headers.ProvidesHeader
+import controllers.helper.GoogleMapsHelper
 import models.user._
 import models.label.{LabelTable, LabelTypeTable}
 import models.label.LabelTable._
@@ -42,7 +43,7 @@ class GalleryController @Inject() (implicit val env: Environment[User, SessionAu
         val labelsShuffled = scala.util.Random.shuffle(labels)
         val jsonList: Seq[JsObject] = labelsShuffled.map(l => Json.obj(
             "label" -> LabelTable.validationLabelMetadataToJson(l),
-            "imageUrl" -> getImageUrl(l.gsvPanoramaId, l.canvasWidth, l.canvasHeight, l.heading, l.pitch, l.zoom)
+            "imageUrl" -> GoogleMapsHelper.getImageUrl(l.gsvPanoramaId, l.canvasWidth, l.canvasHeight, l.heading, l.pitch, l.zoom)
           )
         )
         val labelList: JsObject = Json.obj("labelsOfType" -> jsonList)
@@ -82,7 +83,7 @@ class GalleryController @Inject() (implicit val env: Environment[User, SessionAu
           }
         val jsonList: Seq[JsObject] = labels.map(l => Json.obj(
             "label" -> LabelTable.validationLabelMetadataToJson(l),
-            "imageUrl" -> getImageUrl(l.gsvPanoramaId, l.canvasWidth, l.canvasHeight, l.heading, l.pitch, l.zoom)
+            "imageUrl" -> GoogleMapsHelper.getImageUrl(l.gsvPanoramaId, l.canvasWidth, l.canvasHeight, l.heading, l.pitch, l.zoom)
           )
         )
 
@@ -99,37 +100,5 @@ class GalleryController @Inject() (implicit val env: Environment[User, SessionAu
       )
     }
 
-  }
-
-  /**
-   * Retrieves the static image of the label panorama from the Google Street View Static API. 
-   * Note that this returns the image of the panorama, but doesn't actually include the label. 
-   * More information here: https://developers.google.com/maps/documentation/streetview/intro
-   *
-   * @param label Label to retrieve the static image of.
-   * @return Image URL that represents the background of the label.
-   */
-  def getImageUrl(gsvPanoramaId: String, canvasWidth: Int, canvasHeight: Int, heading: Float, pitch: Float, zoom: Int): String = {
-    val url = "https://maps.googleapis.com/maps/api/streetview?" +
-      "pano=" + gsvPanoramaId +
-      "&size=" + canvasWidth + "x" + canvasHeight +
-      "&heading=" + heading +
-      "&pitch=" + pitch +
-      "&fov=" + getFov(zoom) +
-      "&key=" + Play.configuration.getString("google-maps-api-key").get
-    VersionTable.signUrl(url)
-  }
-
-  /**
-   * Hacky fix to generate the FOV for an image (sort of like zoom).
-   * Determined experimentally.
-   * @param label Label to retrieve the FOV for.
-   */
-  def getFov(zoom: Int): Double = {
-    if (zoom <= 2) {
-      126.5 - zoom * 36.75
-    } else {
-      195.93 / scala.math.pow(1.92, zoom * 1.0)
-    }
   }
 }
